requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Task Scheduling
apscheduler==3.10.4

# Email
secure-smtplib==0.1.1

# HTTP requests
requests==2.31.0

# YAML parsing
pyyaml==6.0.1

# Date/time utilities
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Code execution sandbox
RestrictedPython==7.0

# Progress tracking
tinydb==4.8.0

# Testing framework
pytest==7.4.3
